wip: copy-writing
