+++ i18n/en.json
@@ -737,14 +737,11 @@
   "checksum": "Checksum",
   "choose_matching_people_to_merge": "Choose matching people to merge",
   "city": "City",
-  "cleanup": "Cleanup",
   "cleanup_confirm_description": "You are about to move {count} assets to your device's trash. These assets were created before {date} and are safely backed up on the server. To reclaim storage space, empty your device's trash after this operation",
   "cleanup_deleted_assets": "Moved {count} assets to device trash",
   "cleanup_deleting": "Moving to trash...",
-  "cleanup_description": "Move backed up assets to your device's trash to free up storage. To fully reclaim space, empty your device's trash afterwards. This will not affect files on the server",
   "cleanup_found_assets": "Found {count} backed up assets",
   "cleanup_no_assets_found": "No backed up assets found before this date",
-  "cleanup_settings_subtitle": "Free up device storage",
   "cleanup_step2_description": "Scan your device for assets that have been backed up to the server",
   "cleanup_step3_summary": "{count} assets created before {date} will be moved to your device's trash",
   "clear": "Clear",
@@ -1172,6 +1169,9 @@
   "folders_feature_description": "Browsing the folder view for the photos and videos on the file system",
   "forgot_pin_code_question": "Forgot your PIN?",
   "forward": "Forward",
+  "free_up_space": "Free Up Space",
+  "free_up_space_description": "Move backed-up photos and videos to your device's trash to free up space. Your copies on the server remain safe",
+  "free_up_space_settings_subtitle": "Free up device storage",
   "full_path": "Full path: {path}",
   "gcast_enabled": "Google Cast",
   "gcast_enabled_description": "This feature loads external resources from Google in order to work.",
